requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
PyJWT==2.8.0
cryptography==42.0.7
python-dotenv==1.0.1
jinja2==3.1.4
